Separate out personal data removed spec
The admin_claim_check_spec.rb file has got unwieldy. Let's split it up
into discrete feature spec files where appropriate.
